+++ src/main/kotlin/data/model/GeoResult.kt
@@ -1,0 +1,9 @@
+package org.example.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GeoResult(
+    val latitude: Double,
+    val longitude: Double
+)
