+++ src/main/kotlin/data/model/GeocodingResponse.kt
@@ -1,0 +1,8 @@
+package org.example.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GeocodingResponse(
+    val results: List<GeoResult>?
+)
